Front

+++ server/qraphql/schema/channel/channel.resolvers.js
@@ -1,5 +1,6 @@
 const Channel = require('../../../models/Channel');
 const Message = require('../../../models/Message');
+const User = require('../../../models/User');
 const { extendChannel } = require('../helpers');
 
 const resolvers = {
@@ -8,14 +9,18 @@ const resolvers = {
       const channel = await Channel.findOne({ where: { id } });
       return await extendChannel(channel);
     },
-    channels: async () => {
-      const channels = await Channel.findAll();
+    channels: async (_, __, req) => {
+      if (!req.isAuth || !req.userId) {
+        throw new Error('Authentication failed');
+      }
+      const user = await User.findByPk(req.userId);
+      const channels = await user.getChannels();
       return channels.map(async (c) => await extendChannel(c));
     },
   },
   Mutation: {
     createChannel: async (_, { input }, req) => {
-      if (!req.isAuth || !req.userId || !req.isAdmin) {
+      if (!req.isAuth || !req.userId) {
         throw new Error('Authentication failed');
       }
       const { name, users } = input;
@@ -24,7 +29,7 @@ const resolvers = {
       const channel = await Channel.create({
         name,
       });
-      await channel.addUsers(users);
+      await channel.addUsers([...users, req.userId]);
       return await extendChannel(channel);
     },
     deleteChannel: async (_, { id }, req) => {
@@ -34,6 +39,9 @@ const resolvers = {
       const channel = await Channel.findByPk(id);
       if (!channel) {
         throw new Error('Channel not found');
+      }
+      if (channel.name === 'general') {
+        throw new Error('Cannot delete the default channel');
       }
       await Message.destroy({ where: { channel: id } });
       await channel.destroy();
